+++ app/src/main/java/com/example/android/cloudy/fragments/CurrentForecastFragment.java
@@ -1,32 +1,197 @@
 package com.example.android.cloudy.fragments;
 
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
+import android.support.v7.widget.CardView;
+import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
 
 import com.example.android.cloudy.R;
+import com.example.android.cloudy.data.model.remote.CollectWeatherData;
+import com.example.android.cloudy.data.model.remote.WeatherCallback;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.common.api.Status;
+import com.google.android.gms.location.places.Place;
+import com.google.android.gms.location.places.Places;
+import com.google.android.gms.location.places.ui.PlaceAutocompleteFragment;
+import com.google.android.gms.location.places.ui.PlaceSelectionListener;
+
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
+import static android.R.color.black;
 
 /**
  * Created by ccu17 on 26/04/2017.
  */
 
-public class CurrentForecastFragment extends Fragment {
+public class CurrentForecastFragment extends Fragment implements GoogleApiClient.OnConnectionFailedListener {
+
+    @BindView(R.id.time)
+    TextView TimeText;
+    @BindView(R.id.day)
+    TextView CurrentDay;
+    @BindView(R.id.forecast)
+    TextView weatherForecast;
+    @BindView(R.id.location)
+    TextView chosenLocation;
+    @BindView(R.id.current_forecast_card_view)
+    CardView forecastCardView;
+    @BindView(R.id.current_weather_icon)
+    ImageView currentWeatherIcon;
+
+    private CollectWeatherData collectWeatherData = new CollectWeatherData();
+    public PlaceAutocompleteFragment autocompleteFragment;
+    public Menu menuOptions;
+    public GoogleApiClient googleApiClient;
+
+    public String selectedPlace;
 
     public CurrentForecastFragment() {
         // Required empty public constructor
     }
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    public static CurrentForecastFragment newInstance() {
+        CurrentForecastFragment fragment = new CurrentForecastFragment();
+        return fragment;
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_current_forecast, container, false);
+        View view = inflater.inflate(R.layout.fragment_current_forecast, container, false);
+        ButterKnife.bind(this, view);
+
+        setCurrentTime();
+        setCurrentDay();
+        googleApiInit();
+
+        autocompleteFragment = (PlaceAutocompleteFragment) getActivity().getFragmentManager().findFragmentById(R.id.place_autocomplete_fragment);
+
+        autocompleteFragment.setOnPlaceSelectedListener(new PlaceSelectionListener() {
+            @Override
+            public void onPlaceSelected(Place place) {
+                selectedPlace = place.getName().toString();
+                chosenLocation.setText(place.getName().toString());
+                collectCurrentWeatherData();
+            }
+
+            @Override
+            public void onError(Status status) {
+                Log.i("CHRIS", "An error occurred: " + status);
+            }
+        });
+
+        return view;
+    }
+
+    @Override
+    public void onStart() {
+        googleApiClient.connect();
+        super.onStart();
+    }
+
+
+    @Override
+    public void onStop() {
+        googleApiClient.disconnect();
+        super.onStop();
+    }
+
+    public void googleApiInit() {
+        googleApiClient = new GoogleApiClient
+                .Builder(this.getActivity())
+                .addApi(Places.GEO_DATA_API)
+                .addApi(Places.PLACE_DETECTION_API)
+                .enableAutoManage((FragmentActivity) this.getActivity(), this)
+                .build();
+    }
+
+    public void setCurrentTime() {
+        Calendar c = Calendar.getInstance();
+        SimpleDateFormat df = new SimpleDateFormat("HH:mm");
+        String currentDateTimeString = df.format(c.getTime());
+        TimeText.setText(currentDateTimeString);
+    }
+
+
+    public void setCurrentDay() {
+        Calendar c = Calendar.getInstance();
+        SimpleDateFormat df = new SimpleDateFormat("EEEE");
+        String currentDateTimeString = df.format(c.getTime());
+        CurrentDay.setText(currentDateTimeString);
+    }
+
+    public void collectCurrentWeatherData() {
+        collectWeatherData.collectWeather(selectedPlace , new WeatherCallback() {
+            @Override
+            public void success(String description, double tempMin, double tempMax, double windInMph) {
+                int minTempInCelcious = (int) (tempMin - 273.15);
+                int maxTempInCelcious = (int) (tempMax - 273.15);
+                
+                switch (description) {
+                    case "clear sky":
+                        currentWeatherIcon.setImageResource(R.drawable.sunny);
+                        break;
+                    case "mist":
+                        currentWeatherIcon.setImageResource(R.drawable.mist);
+                        break;
+                    case "few clouds":
+                        currentWeatherIcon.setImageResource(R.drawable.cloudy);
+                        break;
+                    case "scattered clouds":
+                        currentWeatherIcon.setImageResource(R.drawable.cloudy);
+                        break;
+                    case "overcast clouds":
+                        currentWeatherIcon.setImageResource(R.drawable.cloudy);
+                        break;
+                    case "heavy intensity rain":
+                        currentWeatherIcon.setImageResource(R.drawable.rain);
+                        break;
+                    case "light intensity shower rain":
+                        currentWeatherIcon.setImageResource(R.drawable.rain);
+                        break;
+                    case "shower rain":
+                        currentWeatherIcon.setImageResource(R.drawable.rain);
+                        break;
+                    case "light rain":
+                        currentWeatherIcon.setImageResource(R.drawable.rain);
+                        break;
+                    default:
+                        currentWeatherIcon.setImageResource(R.drawable.sunny);
+                        break;
+                }
+
+                weatherForecast.setTextColor(getResources().getColor(black));
+                weatherForecast.setText("The Weather forecast for today is " + description + "."
+                        + "\nThe Minimum temp is " + minTempInCelcious + "°C"
+                        + " & Maximum temp is " + maxTempInCelcious + "°C."
+                        + "\nWind speed is " + windInMph + "mph");
+
+            }
+
+            @Override
+            public void failure(String failed) {
+                Log.i("CHRIS", "Sorry there was an error displaying the weather");
+            }
+        });
+    }
+
+
+    @Override
+    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+
     }
 }
