+++ app/src/main/java/com/example/android/cloudy/fragments/FavouriteFragment.java
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 
 import com.example.android.cloudy.R;
 
+import butterknife.ButterKnife;
+
 /**
  * Created by ccu17 on 26/04/2017.
  */
@@ -27,6 +29,8 @@ public class FavouriteFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favourite_forecasts, container, false);
+        View view = inflater.inflate(R.layout.fragment_favourite_forecasts, container, false);
+        ButterKnife.bind(this, view);
+        return view;
     }
 }
