+++ app/src/main/java/com/example/android/cloudy/fragments/GoogleMapsFragment.java
@@ -8,6 +8,8 @@ import android.view.ViewGroup;
 
 import com.example.android.cloudy.R;
 
+import butterknife.ButterKnife;
+
 /**
  * Created by ccu17 on 26/04/2017.
  */
@@ -27,6 +29,8 @@ public class GoogleMapsFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_google_maps, container, false);
+        View view = inflater.inflate(R.layout.fragment_google_maps, container, false);
+        ButterKnife.bind(this, view);
+        return view;
     }
 }
